# gh-138535: Optimize fill_time for typical timestamps