While file timestamps can be anything the file system can store, most lie between the recent past and the near future.  Optimize fill_time for typical timestamps in three ways:

- When possible, convert to nanoseconds with C arithmetic.
- When using C arithmetic and the seconds member is not required (for st_birthtime), avoid creating a long object.
- When using C arithmetic, reorder the code to avoid the null checks implied in Py_XDECREF.

This improves `python -m pyperf timeit -s 'import os' 'os.stat(".")'` from `1.26 us +- 0.02 us` to `1.15 us +- 0.01 us` on Linux 6.16.2.arch1-1 btrfs and `--enable-optimizations --with-lto`.

I found this while [implementing `os.stat_result.st_birthtime` on Linux](https://github.com/python/cpython/pull/136334) and trying not to regress performance.  This is a small change that should be an improvement on all platforms, so I've submitted it separately.

This needs tests, presumably in `test_os.py` `UtimeTests.test_large_time`, though it's actually testing `os.stat`.  But that test currently only runs on NTFS on Windows.  I'll look at Linux filesystem support tomorrow.

<!-- gh-issue-number: gh-138535 -->
* Issue: gh-138535
<!-- /gh-issue-number -->