@jbosboom 2025-09-06
I want to add tests for timestamps just inside and outside the fast path to `UtimeTests.test_large_time`.  Most but not all Linux filesystems support large timestamps.  The direct way to check the filesystem type is calling `statfs` and comparing the `f_type` member to the magic numbers in `man statfs`.  Unfortunately both `sizeof(struct statfs)` and `sizeof(f_type)` vary across different architectures, which makes it hard to call with ctypes.  Is there precedent for adding a non-public Python interface to a Linux-specific C function just for test purposes?  I didn't see anything about that in the devguide docs.

Another way is to parse `/proc/self/mountinfo` looking for the major:minor of the test file, but parsing that file is nontrivial because it is space-separated with fields that may contain spaces, has optional fields in the middle of lines, and has no documented escaping rules (see `man proc_pid_mountinfo`).