@jbosboom 2025-09-06
I don't think that checking large timestamp support using the filesystem magic number or name is actually that useful.  I think tmpfs has supported 64-bit timestamps as long as the kernel VFS layer has.  btrfs has always had 64-bit timestamps on-disk but I'm not sure they were plumbed in right away.  The in-kernel NTFS driver should also have 64-bit timestamps, but I think most Linux NTFS users use the FUSE implementation ntfs-3g, and that will have the same magic number as other FUSE filesystems.

For the other common Linux filesystems we can't assume large timestamps.  It's still possible to format ext4 with 32-bit timestamps, and anyway ext2 and ext3 use the same `f_type` magic number as ext4.  xfs has only supported 64-bit timestamps since Linux 5.10.

I imagine most programs that care test for large timestamps by seeing if utimes and friends fails or stat returns a clamped timestamp ("Updated file timestamps are set to the greatest value supported by the filesystem that is not greater than the specified time." -- `man utimensat`).  But obviously that can't be part of a test that `os.stat` properly handles large timestamps!

So, reviewers, would you be satisfied just testing this on Windows NTFS by adding a few timestamps to `UtimeTests.test_large_time`?  `fill_time` only depends on `time_t`, and that should be a signed 64-bit type everywhere now.